Implemented pagination
Implemented pagination. This adds two arguments to the `Reactable.Table`
component - `pagination`, a boolean, and `itemsPerPage`, an integer.
Both are optional. If `pagination` is `true`, a `<tr>` with `colspan`
equal to the number of columns in the table and `class="pagination"`
will be rendered after all the content rows with an
`<a class="page-button">` for each page.

From an implementation standpoint this is a pretty naive approach at
the moment, as we're not doing any caching of the children of the table
between state changes.