prototype solution to custom components
Since react will not render child components until the parent component
is rendered (exploratory hypotheis) We need to render the custom
component first and then access the component.  This uses a Component
function of getData().  The getData function could/should be fast
because after the render function all the data could be available via
props or state

I have not tested this with updating data from the custom component.  Or
if the custom component internally changes the data and they table needs
to resort.