Prepend fix for Babel's UMD output to built file
Unlike the commonjs module format, Babel's UMD module output doesn't
include a fix to set the `["default"]` property on a module that wasn't
build with Babel - so it was trying to reference global react
(window.React) as `_react["default"]`. This includes a fix to prepend
`window.React["default"] = window.React` to build/reactable.js. This
isn't perfect (it could potentially interfere with other properties
defined on React) but it is probably the best way to do this short of
writing a plugin to Babel.

 #167