Allow to filter by column: value
This Branch improves the way filterBy workes, by allowing it to be
set with and hash like {'Column': 'search query'} programatically.

Another improvement, that still needs to be touched is the ability to do
the same from the search box in a form of 'Column: search query' that will be
split by the ':' and use each side as column to filter, and value.